# Fixes a bug which causes libc to thrash the futex_wake_all