--- a/lib/wasi/src/syscalls/wasix/futex_wake.rs
+++ b/lib/wasi/src/syscalls/wasix/futex_wake.rs
@@ -33,7 +33,7 @@ pub fn futex_wake<M: MemorySize>(
             }
             true
         } else {
-            false
+            true
         }
     };
     Span::current().record("woken", woken);