@theduke 2023-03-29
But doesn't this break the semantics of the syscall for users?

Because now the user does not have a way to determine if anyone was waiting.