@theduke 2023-03-29
The Linux equivalent returns the number of woken waiters: https://man7.org/linux/man-pages/man2/futex.2.html